Create Yard Greening
Bozhidara has several houses on the Black Sea and wants to landscape the yards of some of them, thereby creating a cozy atmosphere and comfort for his guests. For this purpose, she hired a company.
Write a program that calculates the necessary amount that Bozhidara will have to pay to the company executing the project. The price per square meter is BGN 7.61 including VAT. Because her yard is quite large, the contractor offers an 18% discount off the final price.

Login
Only one line is read from the console:
1. Sq. meters to be landscaped - real number in the range [0.00 … 10000.00].

Exit
Two lines are printed to the console:
• "The final price is: {final price of the service} lv."
• "The discount is: {discount} lv."