Create Deposit Calculator
Write a program that calculates how much you will receive at the end of the deposit period at a given interest rate. Use the following formula: amount = deposited amount + term of the deposit * ((deposited amount * annual interest rate ) / 12).

Login
3 lines are read from the console:
1. Amount deposited – real number in the range [100.00 … 10000.00]
2. Term of the deposit (in months) – an integer in the interval [1…12]
3. Annual interest rate - real number in the interval [0.00 …100.00]

Exit
To print the amount at the end of the term on the console.