Create Vacation Books List
There are a number of books that Joro must read for the summer vacation. Since Joro prefers to play with friends outside, your task is to help him calculate how many hours a day he should devote to reading the necessary literature.

Enter: 3 lines are read from the console
1. Number of pages in the current book - an integer in the range [1…1000].
2. Pages read in 1 hour - an integer in the range [1…1000].
3. The number of days for which he should read the book - an integer in the range [1…1000].

Exit: 
To print to the console the number of hours Joro must spend reading each day.