Create Supplies for School
The school year has already started and the teacher of class 10B - Annie needs to buy a certain number of packets of pens, packets of markers, as well as blackboard cleaner. She is a regular customer of a bookstore, so there is a discount for her, which is some percentage of the total. Write a program that calculates how much money Annie will need to collect to pay the bill, given the following price list:
• Package of chemicals - BGN 5.80.
• Pack of markers - BGN 7.20.
• Detergent - BGN 1.20 (liter)

Enter: 4 lines read from the console
• Number of packages with chemicals - an integer in the range [0...100].
• Number of tag packets - an integer in the interval [0...100].
• Liters of blackboard cleaner - an integer in the range [0…50].
• Percent reduction - an integer in the interval [0...100].

Exit:
To print on the console how much money Annie will need to pay her bill.