Update Repainting
Rumen wants to repaint the living room and has hired craftsmen for the purpose. Write a program that calculates the repair cost at the following prices:

• Safety nylon - BGN 1.50/sq.m
• Paint - BGN 14.50/l
• Paint thinner - BGN 5.00/l

Just in case, to the necessary materials Rumen wants to add another 10% of the amount of paint and 2 sq.m. nylon, of course, and BGN 0.40 for bags. The amount paid to craftsmen for 1 hour of work is equal to 30% of the amount of all material costs.

The input is read from the console and contains exactly 4 lines:
1. Required amount of nylon (in square meters) - an integer in the range [1... 100].
2. Required amount of paint (in liters) - an integer in the interval [1…100].
3. Amount of thinner (in liters) - an integer in the interval [1…30].
4. The hours for which the craftsmen will complete the work - an integer in the interval [1…9].

Exit
To print a single line to the console:
• "{sum of all costs}"